Fuck this, i got assets tho

+++ Source/Level.cpp
@@ -35,6 +35,7 @@ void Level::level_init()
 void Level::media_init()
 {
 	InitAudioDevice();
+
 	menusound = LoadSound("menu.wav");
 	death = LoadSound("death.wav");
 	collectible = LoadSound("collectible.wav");
@@ -220,7 +221,7 @@ void Level::render()
 
 	if (combo_timer > 0)
 	{
-	DrawText(TextFormat("+" "%i",combo_text_value), mario.position.x + 35, mario.position.y, 16, YELLOW); //draw combo text
+	DrawText(TextFormat("+" "%i",combo_text_value), (int)mario.position.x + 35, (int)mario.position.y, 16, YELLOW); //draw combo text
 	}
 	DrawText(TextFormat("%02i",point_count), 210, 10, 24, BLUE);
 }
